requirements.txt
aiodns==1.2.0
aiohttp==3.7.4
altgraph==0.17
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.4.1
astroid==2.4.2
async-timeout==3.0.1
attrs==18.2.0
autopep8==1.6.0
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.7
cloudinary==1.26.0
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
cssselect==1.1.0
distlib==0.3.3
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.2.8
django-cloudinary-storage==0.3.0
django-cors-headers==3.10.0
django-graphql-jwt==0.3.4
django-heroku==0.3.1
django-mdeditor==0.1.18
django-tinymce==3.3.0
fake-useragent==0.1.11
filelock==3.3.2
googletrans==3.0.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.8.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==3.3
iso-639==0.4.5
iso3166==0.9
isodate==0.6.0
isort==5.7.0
lazy-object-proxy==1.4.3
lxml==4.6.2
macholib==1.14
maxminddb==2.2.0
mccabe==0.6.1
multidict==4.5.2
parse==1.19.0
pbr==5.6.0
Pillow==8.4.0
platformdirs==2.4.0
promise==2.3
protobuf==3.17.3
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycares==4.0.0
pycodestyle==2.8.0
pycparser==2.20
pycryptodome==3.11.0
pyee==8.1.0
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
PyJWT==2.3.0
pylint==2.6.0
PyMySQL==1.0.2
pynput==1.7.2
pyobjc-core==7.1
pyobjc-framework-Cocoa==7.1
pyobjc-framework-Quartz==7.1
pyppeteer==0.2.5
pyquery==1.4.3
PySocks==1.6.8
python-decouple==3.5
pytz==2021.3
requests==2.26.0
requests-html==0.10.0
rfc3986==1.5.0
Rx==1.6.1
secure-smtplib==0.1.1
selenium==3.141.0
singledispatch==3.7.0
six==1.16.0
sniffio==1.2.0
soupsieve==1.7.3
sqlparse==0.4.2
static3==0.7.0
stevedore==3.5.0
streamlink==1.0.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.57.0
typing==3.6.6
typing-extensions==3.10.0.2
urllib3==1.26.7
virtualenv==20.9.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
w3lib==1.22.0
webdriver-manager==3.4.2
websocket-client==0.55.0
websockets==8.1
whitenoise==5.3.0
wrapt==1.12.1
xvfbwrapper==0.2.9
yarl==1.3.0
